PP-10038: Add pact tests for invite completion
* These replace some existing selfservice-to-be pact tests (that weren't being verified against adminusers).
* Remove the tests that test error responses as we don't handle different errors in specific ways.
