+++ .secrets.baseline
@@ -76,6 +76,10 @@
       "path": "detect_secrets.filters.allowlist.is_line_allowlisted"
     },
     {
+      "path": "detect_secrets.filters.common.is_baseline_file",
+      "filename": ".secrets.baseline"
+    },
+    {
       "path": "detect_secrets.filters.common.is_ignored_due_to_verification_policies",
       "min_level": 2
     },
@@ -685,27 +689,20 @@
         "line_number": 73
       }
     ],
-    "test/unit/clients/adminusers-client/invite/complete-service-invite.pact.test.js": [
+    "test/unit/clients/adminusers-client/invite/complete-invite.pact.test.js": [
       {
         "type": "Hex High Entropy String",
-        "filename": "test/unit/clients/adminusers-client/invite/complete-service-invite.pact.test.js",
-        "hashed_secret": "614770647df3ab100b871bfc0d20e72c8625a5c4",
-        "is_verified": false,
-        "line_number": 38
-      },
-      {
-        "type": "Hex High Entropy String",
-        "filename": "test/unit/clients/adminusers-client/invite/complete-service-invite.pact.test.js",
+        "filename": "test/unit/clients/adminusers-client/invite/complete-invite.pact.test.js",
         "hashed_secret": "4def56874fa19eaaba377aa6ed3cdf5d70bd2c59",
         "is_verified": false,
         "line_number": 39
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "test/unit/clients/adminusers-client/invite/complete-service-invite.pact.test.js",
+        "filename": "test/unit/clients/adminusers-client/invite/complete-invite.pact.test.js",
         "hashed_secret": "73fab36036653ebf81d45c017039d0be4bc59e64",
         "is_verified": false,
-        "line_number": 40
+        "line_number": 72
       }
     ],
     "test/unit/clients/adminusers-client/invite/complete-user-invite.pact.test.js": [
@@ -874,5 +871,5 @@
       }
     ]
   },
-  "generated_at": "2022-11-10T14:50:10Z"
+  "generated_at": "2022-11-15T14:52:17Z"
 }
